+++ src/FieldDefinition/References.php
@@ -35,16 +35,31 @@ final class References implements Resolvable
     private Count $count;
 
     /**
+     * @phpstan-var array<string, \Closure|mixed|Resolvable>
+     *
+     * @psalm-var array<string, \Closure|mixed|Resolvable>
+     *
+     * @var array<string, \Closure|mixed|Resolvable>
+     */
+    private array $fieldDefinitionOverrides;
+
+    /**
      * @phpstan-param class-string<T> $className
+     * @phpstan-param array<string, \Closure|mixed|Resolvable> $fieldDefinitionOverrides
      *
      * @psalm-param class-string<T> $className
+     * @psalm-param array<string, \Closure|mixed|Resolvable> $fieldDefinitionOverrides
+     *
+     * @param array<string, \Closure|mixed|Resolvable> $fieldDefinitionOverrides
      */
     public function __construct(
         string $className,
-        Count $count
+        Count $count,
+        array $fieldDefinitionOverrides = []
     ) {
         $this->className = $className;
         $this->count = $count;
+        $this->fieldDefinitionOverrides = $fieldDefinitionOverrides;
     }
 
     /**
@@ -61,6 +76,7 @@ final class References implements Resolvable
         return $fixtureFactory->createMany(
             $this->className,
             $this->count,
+            $this->fieldDefinitionOverrides,
         );
     }
 }